Refactoring Verification endpoints and preparing MultiEndorser (#11)
* Refactoring Verification endpoints and preparing MultiEndorser


* Verification Cleanup into Verification State


* Update and verify test case block and proof


* Remove incorrect imports


* address warnings

* further cleanup

* cleanup

* unbreak build; fix code

* simplify verification library

* cleanup test

* cargo fmt

+++ Nimble/src/coordinator.rs
@@ -208,8 +208,11 @@ impl Call for CoordinatorState {
 
     {
       let meta_block = {
-        let metadata =
-          helper::pack_metadata_information(tail_hash.clone(), hash_of_block, ledger_height);
+        let metadata = helper::pack_metadata_information(
+          tail_hash.clone(),
+          hash_of_block,
+          ledger_height as usize,
+        );
         let signatures = vec![signature.clone()];
         MetaBlock {
           metadata,
